+++ kotlin-js-action-plugin/src/main/kotlin/com/rnett/action/AutoBuildWorkflow.kt
@@ -1,24 +1,55 @@
 package com.rnett.action
 
 import org.gradle.api.Project
+import org.gradle.api.Task
 import java.io.File
 
+/**
+ * Create a task that calls [generateAutoBuildWorkflow].
+ *
+ * @param add the arguments to `git add`.  `dist` by default.
+ * @param message the commit message.
+ * @param javaVersion the java version to use in the workflow (necessary for gradle).
+ * @param gradleTasks the gradle command line.  Called like `./gradlew $gradleTasks`.
+ * @param runner the github actions runner to use.
+ * @param dir the directory to create `.github/workflows/auto_build_action.yml` in.  `rootDir` by default.
+ */
+fun Project.autoBuildWorkflowTask(
+    add: String = "dist",
+    message: String = "Commit new $add",
+    gradleTasks: String = "build",
+    javaVersion: String = "15",
+    runner: String = "ubuntu-latest",
+    dir: File = rootDir,
+    configure: Task.() -> Unit = {}
+) = tasks.register("generateAutoBuildWorkflow") {
+    doLast {
+        generateAutoBuildWorkflow(dir, add, message, gradleTasks, javaVersion, runner)
+    }
+    configure()
+}
 
 /**
  * Generate sa GitHub action workflow to build (with `gradlew assemble`) and commit `dist/` on each push, if the updates one wasn't build locally.
  *
- * Generates the workflow file in `$rootDir/.github/workflows/`.
+ * Generates the workflow file in `$rootDir/.github/workflows/` by default.
  *
  * @param add the arguments to `git add`.  `dist` by default.
  * @param message the commit message.
- * @param javaVersion The java version to use in the workflow (necessary for gradle).
+ * @param javaVersion the java version to use in the workflow (necessary for gradle).
+ * @param gradleTasks the gradle command line.  Called like `./gradlew $gradleTasks`.
+ * @param runner the github actions runner to use.
+ * @param dir the directory to create `.github/workflows/auto_build_action.yml` in.  `rootDir` by default.
  */
 fun Project.generateAutoBuildWorkflow(
     add: String = "dist",
     message: String = "Commit new $add",
-    javaVersion: String = "15"
+    gradleTasks: String = "build",
+    javaVersion: String = "15",
+    runner: String = "ubuntu-latest",
+    dir: File = rootDir
 ) {
-    com.rnett.action.generateAutoBuildWorkflow(rootDir, add, message, javaVersion)
+    generateAutoBuildWorkflow(dir, add, message, gradleTasks, javaVersion, runner)
 }
 
 /**
@@ -27,13 +58,17 @@ fun Project.generateAutoBuildWorkflow(
  * @param dir the directory to create `.github/workflows/auto_build_action.yml` in.
  * @param add the arguments to `git add`.  `dist` by default.
  * @param message the commit message.
- * @param javaVersion The java version to use in the workflow (necessary for gradle).
+ * @param javaVersion the java version to use in the workflow (necessary for gradle).
+ * @param gradleTasks the gradle command line.  Called like `./gradlew $gradleTasks`.
+ * @param runner the github actions runner to use.
  */
 fun generateAutoBuildWorkflow(
     dir: File,
     add: String = "dist",
     message: String = "Commit new $add",
-    javaVersion: String = "15"
+    gradleTasks: String = "build",
+    javaVersion: String = "15",
+    runner: String = "ubuntu-latest"
 ) {
     val file = File("$dir/.github/workflows/auto_build_action.yml")
     file.parentFile.apply {
@@ -55,7 +90,7 @@ on:
 jobs:
   build:
     name: Update dist
-    runs-on: ubuntu-latest
+    runs-on: $runner
 
     steps:
       - uses: actions/checkout@v2
@@ -69,7 +104,7 @@ jobs:
         run: chmod +x gradlew
         
       - name: Build
-        run: ./gradlew assemble
+        run: ./gradlew $gradleTasks
         
       - name: Commit dist
         uses: EndBug/add-and-commit@v7